refactor: jwt 토큰을 사용하여 멤버를 찾는 기능 리팩토링
