+++ backend/src/main/java/mouda/backend/auth/business/AuthService.java
@@ -9,9 +9,10 @@ import lombok.RequiredArgsConstructor;
 import mouda.backend.auth.Infrastructure.KakaoOauthClient;
 import mouda.backend.auth.exception.AuthErrorMessage;
 import mouda.backend.auth.exception.AuthException;
+import mouda.backend.auth.implement.JwtProvider;
+import mouda.backend.auth.implement.MemberFinder;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
-import mouda.backend.auth.util.JwtProvider;
 import mouda.backend.auth.util.TokenDecoder;
 import mouda.backend.darakbang.infrastructure.DarakbangRepository;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
@@ -28,6 +29,7 @@ public class AuthService {
 	private final DarakbangRepository darakbangRepository;
 	private final DarakbangMemberRepository darakbangMemberRepository;
 	private final KakaoOauthClient kakaoOauthClient;
+	private final MemberFinder memberFinder;
 
 	public LoginResponse oauthLogin(OauthRequest oauthRequest) {
 		String kakaoIdToken = kakaoOauthClient.getIdToken(oauthRequest.code());
@@ -56,9 +58,7 @@ public class AuthService {
 
 	public Member findMember(String token) {
 		long memberId = jwtProvider.extractMemberId(token);
-		return memberRepository.findById(memberId)
-			.orElseThrow(
-				() -> new AuthException(HttpStatus.UNAUTHORIZED, AuthErrorMessage.UNAUTHORIZED));
+		return memberFinder.find(memberId);
 	}
 
 	public DarakbangMember findDarakbangMember(long darakbangId, Member member) {
